+++ edx_rest_framework_extensions/auth/jwt/authentication.py
@@ -2,7 +2,6 @@
 
 import logging
 
-import jwt
 from django.contrib.auth import get_user_model
 from django.middleware.csrf import CsrfViewMiddleware
 from edx_django_utils.monitoring import set_custom_attribute
@@ -77,10 +76,6 @@ class JwtAuthentication(JSONWebTokenAuthentication):
             # CSRF passed validation with authenticated user
             return user_and_auth
 
-        except jwt.InvalidTokenError as token_error:
-            # Note: I think this case is not used, but will monitor the custom attribute to verify.
-            set_custom_attribute('jwt_auth_failed', 'InvalidTokenError:{}'.format(repr(token_error)))
-            raise exceptions.AuthenticationFailed() from token_error
         except Exception as exception:
             # Errors in production do not need to be logged (as they may be noisy),
             # but debug logging can help quickly resolve issues during development.