+++ edx_rest_framework_extensions/auth/jwt/tests/test_middleware.py
@@ -102,7 +102,7 @@ class TestEnsureJWTAuthSettingsMiddleware(TestCase):
         }
         view_classes = {
             "class_view": SomeClassView,
-            "view_set": views['view_set'].cls,  # pylint: disable=no-member
+            "view_set": views['view_set'].cls,
             "function_view": views['function_view'].view_class,
         }
         view = views[view_type]